refactor: flush TODO: add DevtoolOperator
