+++ packages/akashic-cli-serve/src/client/common/types/plugin.ts
