+++ packages/akashic-cli-serve/src/client/view/molecule/EntityTreeOptionBar.css
@@ -1,5 +1,3 @@
-
-/* TODO use altCSS to integrate this with .instance-toolbar */
 .entity-tree-option-bar {
 	flex: 0 0 auto;
 	display: flex;
